[CITE-174] Addresses PR comments
