+++ citesphere/src/main/java/edu/asu/diging/citesphere/api/v1/user/AddNewItemController.java
@@ -1,5 +1,16 @@
 package edu.asu.diging.citesphere.api.v1.user;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.social.zotero.exception.ZoteroConnectionException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,26 +20,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.Authentication;
-import org.springframework.social.zotero.exception.ZoteroConnectionException;
-import org.springframework.http.ResponseEntity;
-
 import edu.asu.diging.citesphere.api.v1.V1Controller;
-import edu.asu.diging.citesphere.model.bib.ICitation;
-import edu.asu.diging.citesphere.model.bib.IGilesUpload;
-import edu.asu.diging.citesphere.model.bib.impl.Citation;
-import edu.asu.diging.citesphere.user.IUser;
-import edu.asu.diging.citesphere.web.forms.CitationForm;
-import edu.asu.diging.citesphere.core.util.model.ICitationHelper;
-import edu.asu.diging.citesphere.core.util.model.IGilesUtil;
 import edu.asu.diging.citesphere.core.exceptions.CannotFindCitationException;
 import edu.asu.diging.citesphere.core.exceptions.CitationIsOutdatedException;
 import edu.asu.diging.citesphere.core.exceptions.GroupDoesNotExistException;
@@ -37,6 +29,13 @@ import edu.asu.diging.citesphere.core.exceptions.ZoteroItemCreationFailedExcepti
 import edu.asu.diging.citesphere.core.service.ICitationManager;
 import edu.asu.diging.citesphere.core.service.jobs.IUploadFileJobManager;
 import edu.asu.diging.citesphere.core.user.IUserManager;
+import edu.asu.diging.citesphere.core.util.IGilesUtil;
+import edu.asu.diging.citesphere.core.util.model.ICitationHelper;
+import edu.asu.diging.citesphere.model.bib.ICitation;
+import edu.asu.diging.citesphere.model.bib.IGilesUpload;
+import edu.asu.diging.citesphere.model.bib.impl.Citation;
+import edu.asu.diging.citesphere.user.IUser;
+import edu.asu.diging.citesphere.web.forms.CitationForm;
 
 @Controller
 public class AddNewItemController extends V1Controller {
@@ -58,7 +57,7 @@ public class AddNewItemController extends V1Controller {
     @Autowired
     private IUploadFileJobManager jobManager;
 
-    @RequestMapping(value = "/items/create/item/{zoteroGroupId}", method = RequestMethod.POST, consumes = {
+    @RequestMapping(value = "/groups/{groupId}/items/create", method = RequestMethod.POST, consumes = {
         MediaType.MULTIPART_FORM_DATA_VALUE })
     public ResponseEntity<ICitation> createNewItem(Authentication authentication,
             @PathVariable("zoteroGroupId") String zoteroGroupId, @ModelAttribute CitationForm itemWithGiles)
@@ -87,23 +86,29 @@ public class AddNewItemController extends V1Controller {
             try {
                 List<byte[]> fileBytes = new ArrayList<>();
                 gilesUtil.convertFilesToBytesList(fileBytes, itemWithGiles.getFiles());
-                IGilesUpload job;
-                try {
-                    job = jobManager.createGilesJob(user, itemWithGiles.getFiles()[0], fileBytes.get(0), zoteroGroupId,
-                            citation.getKey());
-                } catch (GroupDoesNotExistException e) {
-                    logger.error("Could not create job because group does not exist.", e);
-                    return new ResponseEntity<ICitation>(HttpStatus.BAD_REQUEST);
-                }
                 ObjectMapper mapper = new ObjectMapper();
                 ObjectNode root = mapper.createObjectNode(); 
-                gilesUtil.createJobObjectNode(root, job);
-            } catch (CannotFindCitationException | ZoteroHttpStatusException | ZoteroConnectionException
-                    | CitationIsOutdatedException | ZoteroItemCreationFailedException e) {
+                for (int i=0; i<itemWithGiles.getFiles().length; i++) {
+                    IGilesUpload job;
+                    try {
+                        job = jobManager.createGilesJob(user, itemWithGiles.getFiles()[i], fileBytes.get(i), zoteroGroupId,
+                                citation.getKey());
+                    } catch (GroupDoesNotExistException e) {
+                        logger.error("Could not create job because group does not exist.", e);
+                        return new ResponseEntity<ICitation>(HttpStatus.BAD_REQUEST);
+                    }
+
+                    gilesUtil.createJobObjectNode(root, job);
+                }
+            } catch (CannotFindCitationException | CitationIsOutdatedException e) {
                 return new ResponseEntity<ICitation>(HttpStatus.NOT_FOUND);
+            } catch (ZoteroHttpStatusException | ZoteroConnectionException e) {
+                return new ResponseEntity<ICitation>(HttpStatus.INTERNAL_SERVER_ERROR);
+            } catch (ZoteroItemCreationFailedException e) {
+                return new ResponseEntity<ICitation>(HttpStatus.INTERNAL_SERVER_ERROR);
             }
         }
-        return new ResponseEntity<ICitation>(citation, HttpStatus.OK);
+        return new ResponseEntity<ICitation>(citation, HttpStatus.OK); 
     }
 
 }
