--- citesphere/src/main/java/edu/asu/diging/citesphere/core/service/IPersonalAccessTokenService.java
@@ -1,10 +1,0 @@
-package edu.asu.diging.citesphere.core.service;
-
-import java.util.Optional;
-
-import org.springframework.security.core.Authentication;
-
-public interface IPersonalAccessTokenService {
-
-    public Optional<Authentication> verifyToken(Optional<String> token);
-}
