+++ citesphere/src/main/java/edu/asu/diging/citesphere/core/util/impl/GilesUtil.java
@@ -1,4 +1,4 @@
-package edu.asu.diging.citesphere.core.util.model.impl;
+package edu.asu.diging.citesphere.core.util.impl;
 
 import java.io.IOException;
 import java.util.List;
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 
-import edu.asu.diging.citesphere.core.util.model.IGilesUtil;
+import edu.asu.diging.citesphere.core.util.IGilesUtil;
 import edu.asu.diging.citesphere.model.bib.IGilesUpload;
 
 @Component
