+++ citesphere/src/test/java/edu/asu/diging/citesphere/core/util/model/impl/GilesUtilTest.java
@@ -13,6 +13,7 @@ import org.springframework.mock.web.MockMultipartFile;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
+import edu.asu.diging.citesphere.core.util.impl.GilesUtil;
 import edu.asu.diging.citesphere.model.bib.IGilesUpload;
 import edu.asu.diging.citesphere.model.bib.impl.GilesUpload;
